Fix docs dalapi CodeFactor issues
